requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
plotly>=5.17.0
psutil>=5.9.0
pymupdf>=1.23.0
PyPDF2>=3.0.0
openpyxl>=3.1.0
python-magic-bin>=0.4.27
pyyaml>=6.0
python-Levenshtein>=0.21.0